+++ packages/plugin-backend/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Support for new Backstage backend system
+
 ## [1.0.0] - 2024-04-12
 
 - First release